refactor: extract per-mode validation from CliValidator
Extracts specific logic into StandardModeValidator, LoadfileOnlyValidator,
ProductionSetValidator, and CrossCuttingValidator. Also defers directory
creation to the validation phase and passes pre-parsed delimiters directly
to RequestBuilder.

Closes #470
